enhance deepsearch phase 8
