+++ .env.example
@@ -14,6 +14,8 @@
 # If you use any OpenAI-compatible provider (OpenAI/DeepSeek/etc.), set these:
 OPENAI_API_KEY=
 OPENAI_BASE_URL=https://api.openai.com/v1
+OPENAI_CHAT_MODEL=gpt-4o-mini
+OPENAI_EMBEDDING_MODEL=text-embedding-3-small
 
 # Optional: if chat/embedding/OCR use different provider credentials, set component-specific values.
 CHAT_API_KEY=
@@ -31,6 +33,18 @@ OCR_API_BASE_URL=
 TAVILY_API_KEY=
 
 # ==========================
+# Optional: DeepSearch runtime
+# ==========================
+# Provider routing hint for the planner/external channel (unknown values fall back to `tavily`).
+DEEPSEARCH_WEB_PROVIDER=tavily
+# Plan artifacts directory (JSON plans are persisted here when enabled in DeepSearch config).
+DEEPSEARCH_PLAN_OUTPUT_DIR=./local/deepsearch_runs
+# Tool artifacts directory (per-run trace + tool payload snapshots).
+DEEPSEARCH_TOOL_ARTIFACT_DIR=./local/deepsearch_artifacts
+# External channel cache directory (record/replay/auto).
+DEEPSEARCH_EXTERNAL_CACHE_DIR=./local/deepsearch_artifacts/external_cache
+
+# ==========================
 # Optional: local models (no API)
 # ==========================
 # Chat local (Transformers)
@@ -44,6 +58,7 @@ TAVILY_API_KEY=
 # Embedding local (SentenceTransformers)
 # - Set EMBEDDING_MODEL_PROVIDER=huggingface
 # - Set EMBEDDING_MODEL_NAME to a HuggingFace repo id or a local filesystem path
+# - For OpenAI-compatible embeddings, prefer OPENAI_EMBEDDING_MODEL (it overrides EMBEDDING_MODEL_NAME when provider=openai)
 # EMBEDDING_MODEL_PROVIDER=huggingface
 # EMBEDDING_MODEL_NAME=
 # EMBEDDING_DEVICE=cpu
