+++ application/rag_inference/deepsearch/service_runtime/stages.py
@@ -1,0 +1,35 @@
+import inspect
+import logging
+import time
+from typing import Any, Dict
+
+from core.deepsearch.state import DeepSearchState
+
+logger = logging.getLogger(__name__)
+
+
+class DeepSearchServiceStageMixin:
+    async def _execute_stage(
+        self,
+        label: str,
+        func,
+        *,
+        state: DeepSearchState,
+        stage_timings: Dict[str, Any],
+        **kwargs: Any,
+    ):
+        start = time.perf_counter()
+        try:
+            result = func(**kwargs)
+            if inspect.isawaitable(result):
+                result = await result
+        except Exception as exc:  # pragma: no cover - defensive guardrail
+            logger.exception("DeepSearch stage %s failed: %s", label, exc)
+            state.append_error(str(exc), stage=label)
+            state.mark_failed(label, details={"stage": label, "error": str(exc)})
+            raise
+        elapsed_ms = int((time.perf_counter() - start) * 1000)
+        stage_timings[label] = elapsed_ms
+        state.record_cost(f"stage:{label}", {"latency_ms": elapsed_ms})
+        return result
+
