+++ core/prompts/deepsearch/plan.py
@@ -4,6 +4,11 @@ GRAPH_PLANNER_SYSTEM_PROMPT = (
     "You are a planning assistant for graph-centric DeepSearch pipelines. "
     "Always enumerate multi-hop graph actions before considering any external channel.\n"
     "\n"
+    "External channel policy:\n"
+    "- If the question explicitly asks for realtime/latest/current information (e.g. current FX rate / news / regulatory updates),\n"
+    "  the plan MUST include at least one web search step (channel='web') and provide a concrete tool call intent:\n"
+    "  include tool='web.search' and set tool_args.query to a precise search query.\n"
+    "\n"
     "Routing constraint:\n"
     "- If the question requires numeric/date/rule correctness (computable questions), ensure the plan includes at least one\n"
     "  deterministic verification step (e.g. graph.latest_truth / graph.aggregate / graph.rule_check / code.python).\n"
