+++ encapsulation/database/graph_db/pruned_hipporag_neo4j_cache.py
@@ -3,132 +3,12 @@ from typing import List, Dict, Any, Optional, Set, Tuple
 
 import igraph as ig
 
+from encapsulation.database.graph_db.pruned_hipporag_neo4j_cache_directed import _PrunedHippoRAGNeo4jDirectedCacheMixin
+
 logger = logging.getLogger(__name__)
 
 
-class _PrunedHippoRAGNeo4jCacheMixin:
-    @staticmethod
-    def _neighbor_pair(item: Any) -> Optional[Tuple[str, float]]:
-        if not isinstance(item, tuple) or len(item) < 2:
-            return None
-        node_id = str(item[0] or "").strip()
-        if not node_id:
-            return None
-        try:
-            weight = float(item[1])
-        except (TypeError, ValueError):
-            weight = 1.0
-        return node_id, weight
-
-    @staticmethod
-    def _neighbor_triple(item: Any) -> Optional[Tuple[str, float, str]]:
-        pair = _PrunedHippoRAGNeo4jCacheMixin._neighbor_pair(item)
-        if pair is None:
-            return None
-        node_id, weight = pair
-        relation_type = ""
-        if isinstance(item, tuple) and len(item) >= 3 and item[2] is not None:
-            relation_type = str(item[2])
-        return node_id, weight, relation_type
-
-    @staticmethod
-    def _should_add_reverse_edge_for_predicate(
-        predicate: str,
-        *,
-        direction_policy: str,
-        directed_relations: Set[str],
-        direction_insensitive_relations: Set[str],
-    ) -> bool:
-        """
-        Decide whether to add the reverse edge for an Entity->Entity fact edge.
-
-        - direction_policy=whitelist (legacy): default undirected, except explicitly directed predicates.
-        - direction_policy=blacklist (safer): default directed, except explicitly undirected predicates.
-        """
-        policy = str(direction_policy or "whitelist").strip().lower()
-        pred = str(predicate or "").strip()
-        if not pred:
-            # Fail closed: preserve direction by default.
-            return False
-        if policy == "blacklist":
-            return pred in direction_insensitive_relations
-        return pred not in directed_relations
-
-    def _directed_fact_cache_for_owner(self, owner_key: str) -> Dict[str, List[Tuple[str, float, str]]]:
-        cache = getattr(self, "_directed_fact_cache", None)
-        if cache is None or not isinstance(cache, dict):
-            cache = {}
-            setattr(self, "_directed_fact_cache", cache)
-        owner_cache = cache.get(owner_key)
-        if owner_cache is None or not isinstance(owner_cache, dict):
-            owner_cache = {}
-            cache[owner_key] = owner_cache
-        return owner_cache
-
-    def _load_directed_fact_cache(self, *, owner_id: Optional[Any], force_reload: bool = False) -> None:
-        """
-        Load directed Entity->Entity fact edges (RELATES_TO) into memory for direction-aware PPR.
-
-        This cache complements the existing undirected `_graph_cache`:
-        - `_graph_cache`: undirected adjacency for fast neighbor lookups and push-PPR.
-        - `_directed_fact_cache`: directed adjacency with normalized predicates for direction-aware PPR.
-        """
-        with self.read_lock():
-            if not (getattr(self, "_cache_loaded", False) and getattr(self, "_graph_cache", None)):
-                self._load_graph_cache()
-
-        owner_key = self._owner_key(owner_id) if owner_id is not None else None
-        if owner_key is None:
-            return
-
-        loaded_key = getattr(self, "_directed_fact_cache_loaded_key", None)
-        if loaded_key == owner_key and not force_reload:
-            return
-
-        from core.knowledge_graph.schema import normalize_relation_token
-
-        global_owner = getattr(self, "OWNER_GLOBAL_KEY", "__GLOBAL__")
-        params: Dict[str, Any] = {"global_owner": global_owner, "owner_id": owner_key}
-        rows = self._execute_query(
-            """
-            MATCH (e1:Entity)-[r:RELATES_TO]->(e2:Entity)
-            WHERE COALESCE(e1.owner_id, $global_owner) = $owner_id
-              AND COALESCE(e2.owner_id, $global_owner) = $owner_id
-              AND COALESCE(r.owner_id, $global_owner) = $owner_id
-            RETURN e1.entity_id AS source_id,
-                   e2.entity_id AS target_id,
-                   COALESCE(r.weight, 1.0) AS weight,
-                   r.predicate AS predicate
-            """,
-            params,
-        )
-
-        owner_cache: Dict[str, List[Tuple[str, float, str]]] = {}
-        edge_count = 0
-        for record in rows or []:
-            src = str(record.get("source_id") or "").strip()
-            dst = str(record.get("target_id") or "").strip()
-            if not src or not dst:
-                continue
-            pred = str(record.get("predicate") or "").strip()
-            pred_norm = normalize_relation_token(pred) if pred else ""
-            if not pred_norm:
-                continue
-            try:
-                w = float(record.get("weight") or 1.0)
-            except (TypeError, ValueError):
-                w = 1.0
-            owner_cache.setdefault(src, []).append((dst, w, pred_norm))
-            edge_count += 1
-
-        setattr(self, "_directed_fact_cache_loaded_key", owner_key)
-        cache = getattr(self, "_directed_fact_cache", None)
-        if cache is None or not isinstance(cache, dict):
-            cache = {}
-            setattr(self, "_directed_fact_cache", cache)
-        cache[owner_key] = owner_cache
-        logger.info("Directed fact cache loaded: owner=%s edges=%s", owner_key, edge_count)
-
+class _PrunedHippoRAGNeo4jCacheMixin(_PrunedHippoRAGNeo4jDirectedCacheMixin):
     def extract_subgraph_from_cache_for_ppr_directed(
         self,
         subgraph_node_ids: Set[str],
